Rename the service, guard, interceptor and model folders to plural names
