+++ photo-sharing-application-front-end/src/app/Interceptors/HttpCookieInterceptor.ts
