+++ photo-sharing-application-front-end/src/app/add-image/add-image-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ImageService } from '../service/image/image.service';
+import { ImageService } from '../services/image/image.service';
 import { Router } from '@angular/router';
 import { BehaviorSubject } from 'rxjs';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
