+++ photo-sharing-application-front-end/src/app/guards/email-verification.guard.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from '../service/auth/auth.service';
+import { AuthService } from '../services/auth/auth.service';
 import { Observable, filter, map } from 'rxjs';
 import { toObservable } from '@angular/core/rxjs-interop';
