+++ photo-sharing-application-front-end/src/app/login/login.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { LoginService } from '../service/login/login.service';
+import { LoginService } from '../services/login/login.service';
 import { Router } from '@angular/router';
-import { AuthService } from '../service/auth/auth.service';
-import { CsrfTokenService } from '../service/csrf-token/csrf-token.service';
+import { AuthService } from '../services/auth/auth.service';
+import { CsrfTokenService } from '../services/csrf-token/csrf-token.service';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { BehaviorSubject } from 'rxjs';
