+++ photo-sharing-application-front-end/src/app/models/account.ts
