+++ photo-sharing-application-front-end/src/app/models/comment.ts
