+++ photo-sharing-application-front-end/src/app/models/credentials.ts
