+++ photo-sharing-application-front-end/src/app/models/image-details-post-request.ts
