+++ photo-sharing-application-front-end/src/app/models/registration-form.ts
