+++ photo-sharing-application-front-end/src/app/services/comment/comment.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Page } from '../../model/page';
+import { Page } from '../../models/page';
 
 @Injectable({
   providedIn: 'root',
