+++ photo-sharing-application-front-end/src/app/services/csrf-token/csrf-token.service.ts
