+++ photo-sharing-application-front-end/src/app/services/forgotten-password/forgotten-password.spec.ts
