+++ photo-sharing-application-front-end/src/app/services/image/image.service.spec.ts
