+++ photo-sharing-application-front-end/src/app/services/login/login.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Credentials } from '../../model/credentials';
+import { Credentials } from '../../models/credentials';
 import { Observable } from 'rxjs';
 import { HttpHeaders } from '@angular/common/http';
 import { Buffer } from 'buffer';
