+++ photo-sharing-application-front-end/src/app/services/logout/logout.service.ts
