+++ photo-sharing-application-front-end/src/app/services/registration/registration.service.spec.ts
