+++ photo-sharing-application-front-end/src/app/services/registration/registration.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { RegistrationForm } from '../../model/registration-form';
+import { RegistrationForm } from '../../models/registration-form';
 import { Observable } from 'rxjs';
 
 @Injectable({
