+++ photo-sharing-application-front-end/src/app/services/verification/verification.service.ts
